// templates/io-assist-angular-editable/src/app/io-assist/lib/shared/store/thread/thread.actions.enum.ts
export enum THREAD_ACTIONS {
    FETCH_THREADS = "[Thread] Fetch Threads",
    FETCH_THREADS_SUCCESS = "[Thread] Fetch Threads Success",
    FETCH_THREADS_FAILURE = "[Thread] Fetch Threads Failure",
    RENAME_THREAD = "[Thread] Rename Thread",
    RENAME_THREAD_SUCCESS = "[Thread] Rename Thread Success",
    RENAME_THREAD_FAILURE = "[Thread] Rename Thread Failure",
    DELETE_THREAD = "[Thread] Delete Thread",
    DELETE_THREAD_SUCCESS = "[Thread] Delete Thread Success",
    DELETE_THREAD_FAILURE = "[Thread] Delete Thread Failure",
    CHANGE_ACTIVE_THREAD = "[Thread] Activate Thread",
}